+++ src/training_methods/spd/train_spd.py
@@ -71,7 +71,7 @@ def train(cfg: DictConfig):
         logger=wandb_logger,
         callbacks=[checkpoint_callback, lr_monitor],
         log_every_n_steps=cfg.log_every_n_steps,
-        precision='bf16-mixed',
+        precision='bf16-true',
         benchmark=True,
     )
